+++ src/dawn/native/ProgrammableEncoder.cpp
@@ -61,7 +61,7 @@ namespace dawn::native {
             [&](CommandAllocator* allocator) -> MaybeError {
                 InsertDebugMarkerCmd* cmd =
                     allocator->Allocate<InsertDebugMarkerCmd>(Command::InsertDebugMarker);
-                cmd->length = strlen(groupLabel);
+                cmd->length = static_cast<uint32_t>(strlen(groupLabel));
 
                 char* label = allocator->AllocateData<char>(cmd->length + 1);
                 memcpy(label, groupLabel, cmd->length + 1);
@@ -95,7 +95,7 @@ namespace dawn::native {
             [&](CommandAllocator* allocator) -> MaybeError {
                 PushDebugGroupCmd* cmd =
                     allocator->Allocate<PushDebugGroupCmd>(Command::PushDebugGroup);
-                cmd->length = strlen(groupLabel);
+                cmd->length = static_cast<uint32_t>(strlen(groupLabel));
 
                 char* label = allocator->AllocateData<char>(cmd->length + 1);
                 memcpy(label, groupLabel, cmd->length + 1);
